archive/2025/04/29/projects/tools/exec_env/requirements.txt
# Requirements for Execution Environment
PyYAML
Flask
# Add other dependencies if needed (e.g., pika-client for RabbitMQ, Flask/FastAPI for API) 